Admin Panel Audit & Stability Fixes: Integrated Activation, Announcements, AdBlock & Extensions. Fixed HomeScreen crash and implemented Email Verification backend.

+++ server/app/models/activation_code.py
@@ -4,8 +4,8 @@ from app.database import Base
 
 class ProTransaction(Base):
     __tablename__ = "pro_transactions"
-    id = Column(BigInteger, primary_key=True, autoincrement=True)
-    user_id = Column(BigInteger, ForeignKey("users.id", ondelete="CASCADE"), nullable=False)
+    id = Column(Integer, primary_key=True, autoincrement=True)
+    user_id = Column(Integer, ForeignKey("users.id", ondelete="CASCADE"), nullable=False)
     type = Column(String(50))
     plan_type = Column(String(50))
     activation_code = Column(String(100))
@@ -16,12 +16,12 @@ class ProTransaction(Base):
 
 class ActivationCode(Base):
     __tablename__ = "activation_codes"
-    id = Column(BigInteger, primary_key=True, autoincrement=True)
+    id = Column(Integer, primary_key=True, autoincrement=True)
     code = Column(String(100), unique=True, index=True)
     status = Column(Enum("unused", "used", "disabled", name="code_status_enum"), default="unused")
     plan_type = Column(String(50), default="pro") # pro, premium, etc.
     duration_days = Column(Integer, default=30)
-    used_by = Column(BigInteger, ForeignKey("users.id", ondelete="SET NULL"), nullable=True)
+    used_by = Column(Integer, ForeignKey("users.id", ondelete="SET NULL"), nullable=True)
     batch_id = Column(String(100), index=True, nullable=True)
     created_at = Column(DateTime, default=func.now())
     used_at = Column(DateTime, nullable=True)
